test ok

+++ linux-5.10.54/seti_2/test_seti2.c
@@ -14,6 +14,6 @@ int main (int argc, char *argv[])
     int res1 = __sys_seti_1();
     // Make the syscall to seti_2
    // long res1 = syscall(SYS_seti_1);
-    long res = __sys_seti_2(res1);
+    int res = __sys_seti_2(res1);
     return res;
 }
